+++ packages/plugin/jest.config.js
@@ -1,0 +1,15 @@
+module.exports = {
+  testEnvironment: "node",
+  transform: {
+    "^.+\\.ts?$": [
+      "ts-jest",
+      {
+        tsconfig: "tsconfig.json",
+        isolatedModules: false,
+      },
+    ],
+  },
+  moduleFileExtensions: ["js", "ts", "json"],
+  collectCoverage: true,
+  collectCoverageFrom: ["src/**"],
+};